Feat/s3 aws

+++ internal/constant/model/pagination.go
@@ -3,6 +3,6 @@ package model
 type Pagination struct {
 	CurrentPage  int32 `json:"current_page"`
 	TotalPages   int32 `json:"total_page"`
-	TotalItems   int32 `json:"total_items"`
+	TotalItems   int64 `json:"total_items"`
 	ItemsPerPage int32 `json:"items_per_page"`
 }
